rtc: stm32_rtc: fix fraction difference computation
stm32_rtc_get_second_fraction() returns an unsigned value.
Cast both fractions values (current and reference) before computing
the difference, in order to take negative values into account.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/151340

+++ drivers/st/rtc/stm32_rtc.c
@@ -249,8 +249,8 @@ static uint32_t stm32_rtc_get_second_fraction(struct stm32_rtc_calendar *cal)
 static signed long long stm32_rtc_diff_frac(struct stm32_rtc_calendar *cur,
 					    struct stm32_rtc_calendar *ref)
 {
-	return stm32_rtc_get_second_fraction(cur) -
-		stm32_rtc_get_second_fraction(ref);
+	return (signed long long)stm32_rtc_get_second_fraction(cur) -
+	       (signed long long)stm32_rtc_get_second_fraction(ref);
 }
 
 /*******************************************************************************
